Fix test cases issues, move Heap class to another file
