+++ src/createContext.ts
@@ -3,7 +3,7 @@
 import * as scheme_libs from './alt-langs/scheme/scm-slang/src/stdlib/source-scheme-library'
 import { GLOBAL, JSSLANG_PROPERTIES } from './constants'
 import { call_with_current_continuation } from './cse-machine/continuations'
-import { Heap } from './cse-machine/interpreter'
+import Heap from './cse-machine/heap'
 import * as gpu_lib from './gpu/lib'
 import { AsyncScheduler } from './schedulers'
 import { lazyListPrelude } from './stdlib/lazyList.prelude'
