+++ src/interpreter/interpreter.ts
@@ -4,7 +4,7 @@ import { isEmpty } from 'lodash'
 
 import { UNKNOWN_LOCATION } from '../constants'
 import { LazyBuiltIn } from '../createContext'
-import { Heap } from '../cse-machine/interpreter'
+import Heap from '../cse-machine/heap'
 import { uniqueId } from '../cse-machine/utils'
 import * as errors from '../errors/errors'
 import { RuntimeSourceError } from '../errors/runtimeSourceError'
